+++ README.md
@@ -1,3 +1,1 @@
 # AWS-BE-RS-2022
-
-#SQS
