+++ import-csv-service/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "import-csv-parser",
+  "version": "1.0.0",
+  "description": "",
+  "main": "importFileParser.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "MIT",
+  "dependencies": {
+    "csv-parser": "^3.0.0"
+  }
+}
