--- import-service/package.json
@@ -1,41 +1,0 @@
-{
-  "name": "import-service",
-  "version": "0.0.1",
-  "description": "import service of s3 to bucket",
-  "main": "index.js",
-  "scripts": {
-    "test": "test"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/iZamza/AWS-BE-RS-2022.git"
-  },
-  "keywords": [
-    "aws"
-  ],
-  "author": "Viktar Tsikhmanovich",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/iZamza/AWS-BE-RS-2022/issues"
-  },
-  "homepage": "https://github.com/iZamza/AWS-BE-RS-2022#readme",
-  "dependencies": {
-    "@aws-sdk/client-s3": "^3.121.0",
-    "@babel/plugin-transform-modules-commonjs": "^7.18.6",
-    "aws-sdk": "^2.1167.0",
-    "aws-sdk-client-mock": "^1.0.0",
-    "bluebird": "^3.7.2",
-    "csv-parser": "^3.0.0",
-    "jest": "^28.1.2"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.11.1",
-    "@babel/preset-env": "^7.11.0",
-    "babel-loader": "^8.1.0",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-polyfill": "^6.23.0",
-    "babel-preset-env": "^1.6.0",
-    "serverless-webpack": "^5.3.1",
-    "webpack": "^4.35.2"
-  }
-}
