+++ package.json
@@ -1,6 +1,6 @@
 {
   "dependencies": {
-    "aws-sdk": "^2.1166.0",
+    "aws-sdk": "^2.1171.0",
     "csv-parser": "^3.0.0"
   }
 }
